Separei registro de usuários
antes o registro de usuários era feito por uma adaptação do register do breeze, preferi criar um controller proprio para mexer em usuários terceiros

+++ app/Http/Controllers/ProfileController.php
@@ -9,6 +9,8 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\View\View;
 
+use App\Models\User;
+
 class ProfileController extends Controller
 {
 
@@ -47,21 +49,18 @@ class ProfileController extends Controller
     /**
      * Delete the user's account.
      */
-    public function destroy(Request $request): RedirectResponse
+    public function destroy(Request $request, $id): RedirectResponse
     {
-        $request->validateWithBag('userDeletion', [
-            'password' => ['required', 'current_password'],
-        ]);
-
-        $user = $request->user();
-
-        Auth::logout();
-
+        $user = User::findOrFail($id);
+        if ($request->user()->id === $user->id) {
+            return redirect()
+                ->back()
+                ->with('error', 'Você não pode excluir sua própria conta.');
+        }
         $user->delete();
 
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
-
-        return Redirect::to('/');
+        return redirect()
+            ->route('users.index')
+            ->with('success', 'Usuário excluído com sucesso.');
     }
 }
